requirements.txt
requests==2.23.0
PyMySQL==0.9.3
redis==3.4.1
Django==3.0.5
djangorestframework==3.11.0
